requirements.txt
pydantic~=2.6.1
fastapi~=0.109.2
uvicorn~=0.27.1
python-dotenv~=1.0.1
bcrypt~=4.1.2
motor~=3.3.2